requirements.txt
boto3==1.34.23
datasets==3.1.0
evaluate==0.4.3
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
pdf2image==1.17.0
Pillow==11.0.0
pyperclip==1.9.0
pytesseract==0.3.13
pytest==8.3.3
scikit_learn==1.5.2
torch==2.5.1+cu118
torchvision==0.20.1
transformers==4.47.0.dev0
Werkzeug==3.1.3
